Added ways to change current timepoint displayed
The slider allows the user to drag through timepoints and see how the rendering changes as they drag. There is a lag between the slider and what is being rendered.
The input box reflects whatever timepoint is currently being displayed and the user can also type in the box to change to a certain timepoint.
Both systems don't allow the user to exceed 150 and go below 0.