Display metadata in data table on a pop card with a search bar
Clicking on "Metadata" in the toolbar will bring up a popup card with
all of the data from the params.json file. The variables can be sorted
alphabetically or by it's value. The user can also search for a variable
using the search bar in the header.

Note: Currently params.json is loaded from a local file that was
manually downloaded.