Upload json files corresponding to each plot
In macrophage_states.json, the global id of each macrophage is a key and the
entry contains the timepoint in 'x' and the corresponding state of that
macrophage in 'y'.

In macrophage_summary.json, each state is a key and the entry contains how many
macrophages are in that state at a certain timepoint, where 'x' is the timepoint
and 'y' is the amount.

In spore_aggregate.json, 'x', 'min', 'q1', 'med', 'q3', and 'max' are the keys.
Each number in the array represents the corresponding stat at that timepoint.

In spore_healths.json, the global id of each macrophgae is a key and the entry
contains the timepoint in 'x' and the corresponding health of that spore in 'y'.